+++ src/Websites/Websites/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 * Added support for new access restriction features: ServiceTag, multi-ip and http-headers
+* Added support for Import KeyVault Certificate
+        -New Cmdlet
+		-Import-KeyVaultCertificate.
 
 ## Version 2.0.0
 * Added support for Premium V3 pricing tier